+++ example/lib/src/file_picker_demo.dart
@@ -78,7 +78,7 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
     _resetState();
 
     try {
-      pickedFiles = (await FilePicker.platform.pickFiles(
+      final result = await FilePicker.pickFiles(
         type: _pickingType,
         allowMultiple: _multiPick,
         compressionQuality: 60,
@@ -93,8 +93,9 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
         initialDirectory: _initialDirectoryController.text,
         lockParentWindow: _lockParentWindow,
         withData: true,
-      ))
-          ?.files;
+      );
+      printInDebug("pickedFiles: $result");
+      pickedFiles = result?.files;
       hasUserAborted = pickedFiles == null;
     } on PlatformException catch (e) {
       _logException('Unsupported operation: $e');
@@ -130,8 +131,7 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
     _resetState();
 
     try {
-      pickedFilesAndDirectories =
-          await FilePicker.platform.pickFileAndDirectoryPaths(
+      pickedFilesAndDirectories = await FilePicker.pickFileAndDirectoryPaths(
         type: _pickingType,
         allowedExtensions: (_extension?.isNotEmpty ?? false)
             ? _extension?.replaceAll(' ', '').split(',')
@@ -173,9 +173,10 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
   }
 
   void _clearCachedFiles() async {
+    pickedFiles = [];
     _resetState();
     try {
-      bool? result = await FilePicker.platform.clearTemporaryFiles();
+      bool? result = await FilePicker.clearTemporaryFiles();
       _scaffoldMessengerKey.currentState?.hideCurrentSnackBar();
       _scaffoldMessengerKey.currentState?.showSnackBar(
         SnackBar(
@@ -205,7 +206,7 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
     _resetState();
 
     try {
-      pickedDirectoryPath = await FilePicker.platform.getDirectoryPath(
+      pickedDirectoryPath = await FilePicker.getDirectoryPath(
         dialogTitle: _dialogTitleController.text,
         initialDirectory: _initialDirectoryController.text,
         lockParentWindow: _lockParentWindow,
@@ -239,7 +240,7 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
     _resetState();
 
     try {
-      pickedSaveFilePath = await FilePicker.platform.saveFile(
+      pickedSaveFilePath = await FilePicker.saveFile(
         allowedExtensions: (_extension?.isNotEmpty ?? false)
             ? _extension?.replaceAll(' ', '').split(',')
             : null,
@@ -293,7 +294,7 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
 
     setState(() {
       _isLoading = true;
-      _userAborted = false;
+      _userAborted = true;
     });
   }